refactor: consolidate StyledDetailView into shared sticky module (#1163)
