+++ app/components/Layout/components/Detail/sticky/detailView.styles.ts
@@ -1,5 +1,5 @@
 import styled from "@emotion/styled";
-import { DetailView } from "../../components/Layout/components/Detail/detailView";
+import { DetailView } from "../detailView";
 
 export const StyledDetailView = styled(DetailView)`
   max-height: 100vh;
