--- app/views/SourceStudiesView/sourceStudiesView.styles.ts
@@ -1,6 +1,0 @@
-import styled from "@emotion/styled";
-import { DetailView } from "../../components/Layout/components/Detail/detailView";
-
-export const StyledDetailView = styled(DetailView)`
-  max-height: 100vh;
-`;
